[#31] move: 구매 / 판매 도메인에서 주문 도메인 하나로 이동
