--- src/main/java/bc1/gream/domain/buy/dto/request/BuyBidRequestDto.java
@@ -1,11 +1,0 @@
-package bc1.gream.domain.buy.dto.request;
-
-import lombok.Builder;
-
-@Builder
-public record BuyBidRequestDto(
-    Long price,
-    Integer period
-) {
-
-}
