--- src/main/java/bc1/gream/domain/buy/dto/request/BuyCancelBidRequestDto.java
@@ -1,8 +1,0 @@
-package bc1.gream.domain.buy.dto.request;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties
-public record BuyCancelBidRequestDto() {
-
-}
