--- src/main/java/bc1/gream/domain/buy/dto/response/BuyBidResponseDto.java
@@ -1,8 +1,0 @@
-package bc1.gream.domain.buy.dto.response;
-
-public record BuyBidResponseDto(
-    Long price,
-    Long buyId
-) {
-
-}
