--- src/main/java/bc1/gream/domain/buy/dto/response/BuyCancelBidResponseDto.java
@@ -1,8 +1,0 @@
-package bc1.gream.domain.buy.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties
-public record BuyCancelBidResponseDto() {
-
-}
