--- src/main/java/bc1/gream/domain/buy/repository/BuyRepository.java
@@ -1,8 +1,0 @@
-package bc1.gream.domain.buy.repository;
-
-import bc1.gream.domain.buy.entity.Buy;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BuyRepository extends JpaRepository<Buy, Long> {
-
-}
