+++ src/main/java/bc1/gream/domain/order/dto/request/BuyBidRequestDto.java
@@ -1,0 +1,11 @@
+package bc1.gream.domain.order.dto.request;
+
+import lombok.Builder;
+
+@Builder
+public record BuyBidRequestDto(
+    Long price,
+    Integer period
+) {
+
+}
