+++ src/main/java/bc1/gream/domain/order/dto/request/BuyCancelBidRequestDto.java
@@ -1,0 +1,8 @@
+package bc1.gream.domain.order.dto.request;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties
+public record BuyCancelBidRequestDto() {
+
+}
