+++ src/main/java/bc1/gream/domain/order/dto/response/BuyBidResponseDto.java
@@ -1,0 +1,8 @@
+package bc1.gream.domain.order.dto.response;
+
+public record BuyBidResponseDto(
+    Long price,
+    Long buyId
+) {
+
+}
