+++ src/main/java/bc1/gream/domain/order/dto/response/BuyCancelBidResponseDto.java
@@ -1,0 +1,8 @@
+package bc1.gream.domain.order.dto.response;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties
+public record BuyCancelBidResponseDto() {
+
+}
