+++ src/main/java/bc1/gream/domain/order/dto/response/SellBidResponseDto.java
@@ -1,0 +1,9 @@
+package bc1.gream.domain.order.dto.response;
+
+
+public record SellBidResponseDto(
+    Long price,
+    Long sellId
+) {
+
+}
