+++ src/main/java/bc1/gream/domain/order/repository/SellRepository.java
@@ -1,0 +1,8 @@
+package bc1.gream.domain.order.repository;
+
+import bc1.gream.domain.order.entity.Sell;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SellRepository extends JpaRepository<Sell, Long> {
+
+}
