--- src/main/java/bc1/gream/domain/sell/dto/request/SellCancelBidRequestDto.java
@@ -1,8 +1,0 @@
-package bc1.gream.domain.sell.dto.request;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties
-public record SellCancelBidRequestDto() {
-
-}
