--- src/main/java/bc1/gream/domain/sell/dto/response/SellBidResponseDto.java
@@ -1,9 +1,0 @@
-package bc1.gream.domain.sell.dto.response;
-
-
-public record SellBidResponseDto(
-    Long price,
-    Long sellId
-) {
-
-}
