--- src/main/java/bc1/gream/domain/sell/dto/response/SellCancelBidResponseDto.java
@@ -1,8 +1,0 @@
-package bc1.gream.domain.sell.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties
-public record SellCancelBidResponseDto() {
-
-}
