+++ src/test/java/bc1/gream/test/SellTest.java
@@ -1,9 +1,9 @@
 package bc1.gream.test;
 
-import bc1.gream.domain.sell.entity.Sell;
+import bc1.gream.domain.order.entity.Sell;
 import java.time.LocalDateTime;
 
-public interface SellTest extends UserTest, ProductTest{
+public interface SellTest extends UserTest, ProductTest {
 
     Long TEST_SELL_ID = 1L;
